cree una clase que contiene el envio/recibo de fil
