+++ manifest1.mf
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: Servidor.Servidor
+Main-Class: Redes.Servidor.Servidor
 Class-Path: build/jar
