+++ manifest2.mf
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: Cliente.Cliente
+Main-Class: Redes.Cliente.Cliente
 Class-Path: build/jar
